fix some MSVC warnings
* unreachable code in fz_generate_transition
* signed/unsigned comparison in get_comp_index
* potential dataloss due to conversion from int to short in ucdn_mirror

+++ source/fitz/ucdn.c
@@ -33,7 +33,7 @@ typedef struct {
 } MirrorPair;
 
 typedef struct {
-	int start;
+	unsigned int start;
 	short count, index;
 } Reindex;
 
@@ -207,7 +207,7 @@ unsigned int ucdn_mirror(unsigned int code)
 	if (get_ucd_record(code)->mirrored == 0)
 		return code;
 
-	mp.from = code;
+	mp.from = (unsigned short)code;
 	res = bsearch(&mp, mirror_pairs, BIDI_MIRROR_LEN, sizeof(MirrorPair),
 			compare_mp);
